Old code from Dec/Jan that never got committed.
Was an attempt at finding the original function for every single method
that Udon can call automagically, though it was kinda done without
regard to whether it would be in a useful format for
codegen/interpreting.

I rebased it on top of a couple minor changes I made more recently. Not
sure why I did that at this point, but I did.